add 'echo' range op, for single values
Note: i think storing and returning T by-value is the most correct thing
here.  If there is actually a real rvalue (which might look a bit too
gross for anyone to actually use, eg "echo -(s1 + s2)" ), then you
would risk the original rvalue object being destroyed if stored as
a T && in the range or the iterator, and if returning a T && you
invalidate the item's contents with the move, which would mean you
can't access it more than once.

But anyhow i expect T to be a reference type for probably all real use
cases, or else a trivially-copyable literal.