fix returning reference to local vars
the real problem was that std::declval() returns an rvalue ref,
which means an rvalue ref to the temporary broker object is returned,
which of course is invalid as soon as it is returned.

a slightly smaller fix here would have been to add something like

	template <class T>
	T declval_noref();

and swap out std::declval with declval_noref.

That works, but i kinda liked the idea of specifying the return type
explicitly, thus the first_t helper.